+++ benchmark/startup-worker.ts
@@ -16,6 +16,7 @@ const engineIds = new Set([
   'jsquash',
   'purejsimage',
   'purejsimage-wasm',
+  'purejsimage-experimental-heic',
   'sharp',
   'sharp-single-thread',
 ])